requirements.txt
pandas>=1.5.0
numpy>=1.23.0
requests>=2.28.0
websockets>=10.0
pyyaml>=6.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
asyncio>=3.4.3
aiosqlite>=0.17.0

# Zusätzliche Dependencies für robusteren Bot
urllib3>=1.26.0
certifi>=2022.0.0
cryptography>=3.4.8